restructure into Postprocessor
